DF-11912 fix teamcity builds
Former-commit-id: e99e388e9ec435e984b5fb89aca29a52fd0d78c3

+++ Sources/Internal/PackManager/Private/RequestManager.h
@@ -17,7 +17,7 @@ public:
     void Update();
     bool IsInQueue(const String& packName) const;
     bool Empty() const;
-    unsigned CountRequests() const;
+    size_t CountRequests() const;
     PackRequest& Top();
     PackRequest& Find(const String& packName);
     void Push(const String& packName, float32 priority);
